Move topic route management logic to TopicRouteInfoManager
