+++ broker/src/main/java/org/apache/rocketmq/broker/failover/EscapeBridge.java
@@ -20,31 +20,18 @@ package org.apache.rocketmq.broker.failover;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.Executor;
 import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import java.util.concurrent.ForkJoinPool;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.ThreadFactory;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.locks.Lock;
-import java.util.concurrent.locks.ReentrantLock;
 import java.util.function.Function;
-
 import org.apache.commons.lang3.RandomUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.rocketmq.broker.BrokerController;
 import org.apache.rocketmq.client.consumer.PullResult;
 import org.apache.rocketmq.client.consumer.PullStatus;
 import org.apache.rocketmq.client.exception.MQBrokerException;
-import org.apache.rocketmq.client.impl.factory.MQClientInstance;
 import org.apache.rocketmq.client.impl.producer.TopicPublishInfo;
 import org.apache.rocketmq.client.producer.SendResult;
 import org.apache.rocketmq.client.producer.SendStatus;
@@ -54,14 +41,12 @@ import org.apache.rocketmq.common.constant.LoggerName;
 import org.apache.rocketmq.common.message.MessageConst;
 import org.apache.rocketmq.common.message.MessageDecoder;
 import org.apache.rocketmq.common.message.MessageExt;
+import org.apache.rocketmq.common.message.MessageExtBrokerInner;
 import org.apache.rocketmq.common.message.MessageQueue;
-import org.apache.rocketmq.common.protocol.route.BrokerData;
-import org.apache.rocketmq.common.protocol.route.TopicRouteData;
 import org.apache.rocketmq.logging.InternalLogger;
 import org.apache.rocketmq.logging.InternalLoggerFactory;
 import org.apache.rocketmq.remoting.exception.RemotingException;
 import org.apache.rocketmq.store.GetMessageResult;
-import org.apache.rocketmq.common.message.MessageExtBrokerInner;
 import org.apache.rocketmq.store.MessageStore;
 import org.apache.rocketmq.store.PutMessageResult;
 import org.apache.rocketmq.store.PutMessageStatus;
@@ -69,7 +54,6 @@ import org.apache.rocketmq.store.PutMessageStatus;
 public class EscapeBridge {
     protected static final InternalLogger LOG = InternalLoggerFactory.getLogger(LoggerName.BROKER_LOGGER_NAME);
     private static final long SEND_TIMEOUT = 3000L;
-    private static final long LOCK_TIMEOUT_MILLIS = 3000L;
     private static final int DEFAULT_RETRY_TIMES = 3;
     private static final long DEFAULT_PULL_TIMEOUT_MILLIS = 1000 * 10L;
     private final String innerProducerGroupName;
@@ -77,14 +61,7 @@ public class EscapeBridge {
 
     private final BrokerController brokerController;
 
-    private final ConcurrentMap<String/* topic */, TopicPublishInfo> topicPublishInfoTable = new ConcurrentHashMap<>();
-    private final Lock lockNamesrv = new ReentrantLock();
-    private final ConcurrentMap<String/* Topic */, TopicRouteData> topicRouteTable = new ConcurrentHashMap<>();
-    private final ConcurrentMap<String/* Broker Name */, HashMap<Long/* brokerId */, String/* address */>> brokerAddrTable =
-        new ConcurrentHashMap<>();
     private ExecutorService defaultAsyncSenderExecutor;
-    private ScheduledExecutorService scheduledExecutorService;
-    private static final long POLL_NAMESERVER_INTERVAL = 1000 * 30L;
 
     public EscapeBridge(BrokerController brokerController) {
         this.brokerController = brokerController;
@@ -100,9 +77,6 @@ public class EscapeBridge {
                     new NamedForkJoinWorkerThreadFactory("AsyncEscapeBridgeSender_", brokerController.getBrokerIdentity()),
                     null,
                     false);
-
-                initScheduledTask();
-
                 LOG.info("init executor for escaping messages asynchronously success.");
             } else {
                 throw new RuntimeException("nameserver address is null or empty");
@@ -110,36 +84,7 @@ public class EscapeBridge {
         }
     }
 
-    private void initScheduledTask() {
-        this.scheduledExecutorService = Executors.newSingleThreadScheduledExecutor(new ThreadFactory() {
-            @Override
-            public Thread newThread(Runnable r) {
-                return new Thread(r, "EscapeBridgeScheduledThread");
-            }
-        });
-
-        this.scheduledExecutorService.scheduleAtFixedRate(() -> {
-            try {
-                updateTopicRouteInfoFromNameServer();
-            } catch (Exception e) {
-                LOG.error("ScheduledTask updateTopicRouteInfoFromNameServer exception", e);
-            }
-        }, 10, POLL_NAMESERVER_INTERVAL, TimeUnit.MILLISECONDS);
-    }
-
-    private void updateTopicRouteInfoFromNameServer() {
-        if (null == topicRouteTable || topicRouteTable.isEmpty()) {
-            return;
-        }
-        for (String topic : topicRouteTable.keySet()) {
-            this.updateTopicRouteInfoFromNameServer(topic);
-        }
-    }
-
     public void shutdown() {
-        if (null != this.scheduledExecutorService) {
-            this.scheduledExecutorService.shutdown();
-        }
         if (null != this.defaultAsyncSenderExecutor) {
             this.defaultAsyncSenderExecutor.shutdown();
         }
@@ -168,7 +113,7 @@ public class EscapeBridge {
     }
 
     private SendResult putMessageToRemoteBroker(MessageExtBrokerInner messageExt, final int retryTimes) {
-        final TopicPublishInfo topicPublishInfo = this.tryToFindTopicPublishInfo(messageExt.getTopic());
+        final TopicPublishInfo topicPublishInfo = this.brokerController.getTopicRouteInfoManager().tryToFindTopicPublishInfo(messageExt.getTopic());
         if (null == topicPublishInfo || !topicPublishInfo.ok()) {
             LOG.warn("putMessageToRemoteBroker: no route info of topic {} when escaping message, msgId={}",
                 messageExt.getTopic(), messageExt.getMsgId());
@@ -191,7 +136,7 @@ public class EscapeBridge {
 
             String brokerNameToSend = mq.getBrokerName();
             brokersSent[times] = brokerNameToSend;
-            String brokerAddrToSend = this.findBrokerAddressInPublish(brokerNameToSend);
+            String brokerAddrToSend = this.brokerController.getTopicRouteInfoManager().findBrokerAddressInPublish(brokerNameToSend);
             final SendResult sendResult;
             try {
                 beginTimestampPrev = System.currentTimeMillis();
@@ -231,7 +176,7 @@ public class EscapeBridge {
             try {
                 messageExt.setWaitStoreMsgOK(false);
 
-                TopicPublishInfo topicPublishInfo = this.tryToFindTopicPublishInfo(messageExt.getTopic());
+                TopicPublishInfo topicPublishInfo = this.brokerController.getTopicRouteInfoManager().tryToFindTopicPublishInfo(messageExt.getTopic());
                 final String producerGroup = getProducerGroup(messageExt);
                 final CompletableFuture<SendResult> retryFuture = wrapRetryableFuture(
                     lastBroker -> {
@@ -239,7 +184,7 @@ public class EscapeBridge {
                         messageExt.setQueueId(mqSelected.getQueueId());
 
                         String brokerNameToSend = mqSelected.getBrokerName();
-                        String brokerAddrToSend = this.findBrokerAddressInPublish(brokerNameToSend);
+                        String brokerAddrToSend = this.brokerController.getTopicRouteInfoManager().findBrokerAddressInPublish(brokerNameToSend);
                         return this.brokerController.getBrokerOuterAPI().sendMessageToSpecificBrokerAsync(brokerAddrToSend,
                             brokerNameToSend, messageExt,
                             producerGroup, SEND_TIMEOUT);
@@ -314,105 +259,6 @@ public class EscapeBridge {
         return producerGroup;
     }
 
-    private TopicPublishInfo tryToFindTopicPublishInfo(final String topic) {
-        TopicPublishInfo topicPublishInfo = this.topicPublishInfoTable.get(topic);
-        if (null == topicPublishInfo || !topicPublishInfo.ok()) {
-            this.topicPublishInfoTable.putIfAbsent(topic, new TopicPublishInfo());
-            this.updateTopicRouteInfoFromNameServer(topic);
-            topicPublishInfo = this.topicPublishInfoTable.get(topic);
-        }
-        return topicPublishInfo;
-    }
-
-    private boolean updateTopicRouteInfoFromNameServer(String topic) {
-        try {
-            if (this.lockNamesrv.tryLock(LOCK_TIMEOUT_MILLIS, TimeUnit.MILLISECONDS)) {
-                try {
-                    TopicRouteData topicRouteData = this.brokerController.getBrokerOuterAPI()
-                        .getTopicRouteInfoFromNameServer(topic, SEND_TIMEOUT);
-                    if (topicRouteData != null) {
-                        TopicRouteData old = this.topicRouteTable.get(topic);
-                        boolean changed = this.topicRouteDataIsChange(old, topicRouteData);
-                        if (!changed) {
-                            changed = this.isNeedUpdateTopicRouteInfo(topic);
-                        } else {
-                            LOG.info("the topic[{}] route info changed, old[{}] ,new[{}]", topic, old, topicRouteData);
-                        }
-
-                        if (changed) {
-
-                            for (BrokerData bd : topicRouteData.getBrokerDatas()) {
-                                this.brokerAddrTable.put(bd.getBrokerName(), bd.getBrokerAddrs());
-                            }
-
-                            // Update Pub info
-                            {
-                                TopicPublishInfo publishInfo = MQClientInstance.topicRouteData2TopicPublishInfo(topic, topicRouteData);
-                                publishInfo.setHaveTopicRouterInfo(true);
-                                this.updateTopicPublishInfo(topic, publishInfo);
-
-                            }
-
-                            TopicRouteData cloneTopicRouteData = new TopicRouteData(topicRouteData);
-                            LOG.info("topicRouteTable.put. Topic = {}, TopicRouteData[{}]", topic, cloneTopicRouteData);
-                            this.topicRouteTable.put(topic, cloneTopicRouteData);
-                            return true;
-                        }
-                    } else {
-                        LOG.warn("EscapeBridge: updateTopicRouteInfoFromNameServer, getTopicRouteInfoFromNameServer return null, Topic: {}.", topic);
-                    }
-                } catch (RemotingException e) {
-                    LOG.error("updateTopicRouteInfoFromNameServer Exception", e);
-                } catch (MQBrokerException e) {
-                    LOG.error("updateTopicRouteInfoFromNameServer Exception", e);
-                } finally {
-                    this.lockNamesrv.unlock();
-                }
-            }
-        } catch (InterruptedException e) {
-            LOG.warn("updateTopicRouteInfoFromNameServer Exception", e);
-        }
-        return false;
-    }
-
-    public void updateTopicPublishInfo(final String topic, final TopicPublishInfo info) {
-        if (info != null && topic != null) {
-            TopicPublishInfo prev = this.topicPublishInfoTable.put(topic, info);
-            if (prev != null) {
-                LOG.info("updateTopicPublishInfo prev is not null, " + prev);
-            }
-        }
-    }
-
-    private boolean isNeedUpdateTopicRouteInfo(final String topic) {
-        final TopicPublishInfo prev = this.topicPublishInfoTable.get(topic);
-        return null == prev || !prev.ok();
-    }
-
-    private boolean topicRouteDataIsChange(TopicRouteData olddata, TopicRouteData nowdata) {
-        if (olddata == null || nowdata == null)
-            return true;
-        TopicRouteData old = new TopicRouteData(olddata);
-        TopicRouteData now = new TopicRouteData(nowdata);
-        Collections.sort(old.getQueueDatas());
-        Collections.sort(old.getBrokerDatas());
-        Collections.sort(now.getQueueDatas());
-        Collections.sort(now.getBrokerDatas());
-        return !old.equals(now);
-
-    }
-
-    private String findBrokerAddressInPublish(String brokerName) {
-        if (brokerName == null) {
-            return null;
-        }
-        Map<Long/* brokerId */, String/* address */> map = this.brokerAddrTable.get(brokerName);
-        if (map != null && !map.isEmpty()) {
-            return map.get(MixAll.MASTER_ID);
-        }
-
-        return null;
-    }
 
     public PutMessageResult putMessageToSpecificQueue(MessageExtBrokerInner messageExt) {
         BrokerController masterBroker = this.brokerController.peekMasterBroker();
@@ -423,7 +269,7 @@ public class EscapeBridge {
             try {
                 messageExt.setWaitStoreMsgOK(false);
 
-                final TopicPublishInfo topicPublishInfo = this.tryToFindTopicPublishInfo(messageExt.getTopic());
+                final TopicPublishInfo topicPublishInfo = this.brokerController.getTopicRouteInfoManager().tryToFindTopicPublishInfo(messageExt.getTopic());
                 List<MessageQueue> mqs = topicPublishInfo.getMessageQueueList();
 
                 if (null == mqs || mqs.isEmpty()) {
@@ -437,7 +283,7 @@ public class EscapeBridge {
                 messageExt.setQueueId(mq.getQueueId());
 
                 String brokerNameToSend = mq.getBrokerName();
-                String brokerAddrToSend = this.findBrokerAddressInPublish(brokerNameToSend);
+                String brokerAddrToSend = this.brokerController.getTopicRouteInfoManager().findBrokerAddressInPublish(brokerNameToSend);
                 final SendResult sendResult = this.brokerController.getBrokerOuterAPI().sendMessageToSpecificBroker(
                     brokerAddrToSend, brokerNameToSend,
                     messageExt, this.getProducerGroup(messageExt), SEND_TIMEOUT);
@@ -522,10 +368,10 @@ public class EscapeBridge {
 
     protected MessageExt getMessageFromRemote(String topic, long offset, int queueId, String brokerName) {
         try {
-            String brokerAddr = this.findBrokerAddressInSubscribe(brokerName, MixAll.MASTER_ID, false);
+            String brokerAddr = this.brokerController.getTopicRouteInfoManager().findBrokerAddressInSubscribe(brokerName, MixAll.MASTER_ID, false);
             if (null == brokerAddr) {
-                this.updateTopicRouteInfoFromNameServer(topic);
-                brokerAddr = this.findBrokerAddressInSubscribe(brokerName, MixAll.MASTER_ID, false);
+                this.brokerController.getTopicRouteInfoManager().updateTopicRouteInfoFromNameServer(topic, true, false);
+                brokerAddr = this.brokerController.getTopicRouteInfoManager().findBrokerAddressInSubscribe(brokerName, MixAll.MASTER_ID, false);
 
                 if (null == brokerAddr) {
                     LOG.warn("can't find broker address for topic {}", topic);
@@ -546,36 +392,4 @@ public class EscapeBridge {
         return null;
     }
 
-    private String findBrokerAddressInSubscribe(
-        final String brokerName,
-        final long brokerId,
-        final boolean onlyThisBroker
-    ) {
-        if (brokerName == null) {
-            return null;
-        }
-        String brokerAddr = null;
-        boolean found = false;
-
-        Map<Long/* brokerId */, String/* address */> map = this.brokerAddrTable.get(brokerName);
-        if (map != null && !map.isEmpty()) {
-            brokerAddr = map.get(brokerId);
-            boolean slave = brokerId != MixAll.MASTER_ID;
-            found = brokerAddr != null;
-
-            if (!found && slave) {
-                brokerAddr = map.get(brokerId + 1);
-                found = brokerAddr != null;
-            }
-
-            if (!found && !onlyThisBroker) {
-                Map.Entry<Long, String> entry = map.entrySet().iterator().next();
-                brokerAddr = entry.getValue();
-                found = true;
-            }
-        }
-
-        return brokerAddr;
-
-    }
 }
